firmware_unificado/configurator/requirements.txt
PySide6
cantools
pyserial
